+++ src/main/java/com/imsweb/algorithms/internal/CensusData.java
@@ -53,9 +53,16 @@ public class CensusData {
     private String _acsPctPov1418OtherMulti;
     private String _acsPctPov1418White;
     private String _acsPctPov1418WhiteNonHisp;
-    
+
     // Cancer Reporting Zone
     private String _cancerReportingZone;
+
+    // EPHT SubCounty
+    private String _epht2010GeoId5k;
+    private String _epht2010GeoId20k;
+
+    // TractEstCongressDist
+    private String _tractEstCongressDist;
 
     public String getAcsPctPov0610AIAN() {
         return _acsPctPov0610AIAN;
@@ -351,6 +358,30 @@ public class CensusData {
 
     public void setYostQuintile1418US(String str) {
         _yostQuintile1418US = str;
+    }
+
+    public String getEpht2010GeoId20k() {
+        return _epht2010GeoId20k;
+    }
+
+    public void setEpht2010GeoId20k(String epht2010GeoId20k) {
+        _epht2010GeoId20k = epht2010GeoId20k;
+    }
+
+    public String getEpht2010GeoId5k() {
+        return _epht2010GeoId5k;
+    }
+
+    public void setEpht2010GeoId5k(String epht2010GeoId5k) {
+        _epht2010GeoId5k = epht2010GeoId5k;
+    }
+
+    public String getTractEstCongressDist() {
+        return _tractEstCongressDist;
+    }
+
+    public void setTractEstCongressDist(String tractEstCongressDist) {
+        _tractEstCongressDist = tractEstCongressDist;
     }
 
     public String getCancerReportingZone() {